✨ feat: count Integer -> Long type으로 수정

+++ src/main/java/com/mulmeong/comment/read/entity/ShortsRecomment.java
@@ -23,6 +23,6 @@ public class ShortsRecomment {
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer likeCount;
-    private Integer dislikeCount;
+    private Long likeCount;
+    private Long dislikeCount;
 }
