+++ src/main/java/com/mulmeong/comment/read/vo/FeedRecommentResponseVo.java
@@ -15,6 +15,6 @@ public class FeedRecommentResponseVo {
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer likeCount;
-    private Integer dislikeCount;
+    private Long likeCount;
+    private Long dislikeCount;
 }
