+++ src/main/java/com/mulmeong/comment/read/vo/ShortsCommentResponseVo.java
@@ -15,8 +15,8 @@ public class ShortsCommentResponseVo {
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer likeCount;
-    private Integer dislikeCount;
-    private Integer recommentCount;
+    private Long likeCount;
+    private Long dislikeCount;
+    private Long recommentCount;
     private boolean isDeleted;
 }
